fix: resolve circular dependency and RatePriority undefined crash
